requirements.txt
yfinance
psycopg2-binary
fastapi
uvicorn